requirements-dev.txt
ruff==0.8.4
black==24.10.0
mypy==1.13.0
django-stubs==5.1.1
